Bugfix: Disabling default tab icons crashes Safari
This was caused by an infinite loop in the TitleWatcher. Since the value
of `emoji` on some websites was null, whenever we checked for the
presence of `emoji` in the document title, it was never found. As a
result, the extension would continually update the title, which
would activate the TitleWatcher, which would update the title, etc.